services/eval-runner/pyproject.toml
[project]
name = "eval-runner"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "pandas",
  "pyyaml",
  "tqdm",
  "boto3",
  "requests",
  "python-dotenv",
]

[tool.uv]
dev-dependencies = ["pytest", "ruff"]
